@@ -3,45 +3,33 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear",
-    "validate": "svelte-check",
-    "test": "jest"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "test": "vitest"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-node-resolve": "^15.1.0",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-typescript": "^11.1.3",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/svelte": "^3.2.2",
-    "@tsconfig/svelte": "^3.0.0",
-    "@types/node": "^16.18.48",
-    "jest": "^29.6.4",
-    "jest-environment-jsdom": "^29.6.4",
-    "rollup": "^2.79.1",
-    "rollup-plugin-css-only": "^4.3.0",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-svelte": "^7.1.6",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.59.2",
+    "@sveltejs/vite-plugin-svelte": "^2.4.5",
+    "@testing-library/svelte": "^4.0.3",
+    "@tsconfig/svelte": "^5.0.2",
+    "svelte": "^4.2.0",
     "svelte-check": "^3.5.0",
-    "svelte-jester": "^2.3.2",
     "svelte-preprocess": "^5.0.4",
     "svelte-spa-router": "^3.3.0",
-    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tslib": "^2.6.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9",
+    "vitest": "^0.34.3"
   },
   "dependencies": {
     "dayjs": "^1.11.9",
     "dotenv": "^16.3.1",
-    "firebase": "^10.3.1",
-    "sirv-cli": "^2.0.2"
+    "firebase": "^10.3.1"
   },
   "engines": {
     "node": ">=16 <17"
-  }
+  },
+  "type": "module"
 }
