@@ -1,5 +1,5 @@
 import { initializeApp } from "firebase/app";
-import App from './App.svelte';
+import App from "./App.svelte";
 
 document.oncontextmenu = () => {
   alert("\t※※※注意※※※\n★★★右クリック禁止！★★★");
@@ -22,12 +22,12 @@ if ("serviceWorker" in navigator) {
 
 // Firebase の設定
 const config = {
-  databaseURL: process.env.DATABASE_URL,
+  databaseURL: (import.meta as any).env.VITE_DATABASE_URL,
 };
 initializeApp(config);
 
 const app = new App({
-  target: document.body,
+  target: document.getElementById("app")!,
 });
 
 export default app;
