@@ -1,5 +1,14 @@
-import { get, getDatabase, limitToLast, onValue, orderByChild, push, query, ref } from "firebase/database";
+import {
+  getDatabase,
+  limitToLast,
+  onValue,
+  orderByChild,
+  push,
+  query,
+  ref,
+} from "firebase/database";
 import type { DatabaseReference } from "firebase/database";
+import type { Message } from "../types/message";
 
 export default class MessageRepository {
   private ref: DatabaseReference;
@@ -8,18 +17,26 @@ export default class MessageRepository {
     this.ref = ref(getDatabase(), "messages");
   }
 
-  broadcast(callback: Function) {
-    onValue(query(this.ref, orderByChild("sortKey"), limitToLast(20)), (snapshot) => {
-      const r = [];
-      snapshot.forEach((c) => {
-        r.push(c.val());
-      });
-      callback(r.slice());
-    });
+  broadcast(callback: (message: Message[]) => void) {
+    onValue(
+      query(this.ref, orderByChild("sortKey"), limitToLast(20)),
+      (snapshot) => {
+        const r = [] as Message[];
+        snapshot.forEach((c) => {
+          r.push(c.val());
+        });
+        callback(r.slice());
+      }
+    );
   }
 
   async create(username: string, content: string) {
     const timestamp = Date.now();
-    return push(this.ref, { username, content, timestamp, sortKey: timestamp * -1 });
+    return push(this.ref, {
+      username,
+      content,
+      timestamp,
+      sortKey: timestamp * -1,
+    });
   }
 }
